+++ src/filter_stack_trace.ts
